Build f4pga-arch-defs from source
Unfortunately I've run into an issue where VPR seems to segfault while
building f4pga-arch-defs on macOS if TBB is enabled; so I've had to
disable it when building f4pga-arch-defs on macOS. However, this is also
the first instance I've found where TBB actually has an (extremely
significant!) effect on VPR's parallelism, so it runs way slower as a
result, and so it's faster to build it inside a Linux VM instead. As a
result I ended up patching some other things to make them work on Linux.