Actually get f4pga-arch-defs to build in linux-builder
For some reason the regression tests seem to freeze on Linux; it's
probably fixable, but running them isn't very important anyway.

I also had to increase the Linux builder's resources, and it turned out
that my truncated commit hash for f4pga-arch-defs had been broken this
whole time; but for some reason it only resulted in the output being
wrong on Linux? Weird.