vivado: use less hacks
The installation process could now be replicated without any doctoring of files:
it now behaves the same as if you did a 'Download and Install Later', installed
a base Vivado install using that, and then added on the rest of the modules you
wanted (except that we use `PAYLOAD_LOCATION_FROM_USER`).

I made it a 2-stage installation for the same reason I made the previous hacky
solution: to avoid having to re-install the whole thing to change the much
smaller ancillary modules, and to allow sharing the bulk of the install between
your system and Nix builds.

I'm thinking of moving some of this logic from `vivado` into `xinstall`, since
most of it could be used to install Petalinux or a standalone DocNav just as
easily as Vivado.