fex: 2407 -> 2408
I also removed the cross-compiled glibc headers for aarch64, to fix the
weird issue where for some reason a version built in isolation would end
up using a natively-compiled glibc, and a version built as part of a
NixOS config would use a cross-compiled one (which also then had to be
built from source, since the binary cache doesn't contain cross-compiled
packages from a platform to itself).

FEX can only be run (and hence tested) on aarch64 anyway so we can just
rely on the native ones.