Update VTR to latest version
Specifically, the last merge commit with passing CI.

I'm not entirely sure what the consequences are of marking all segments
as 'general'; is it that VPR previously treated all routes that way, so
it makes no difference, or is it that it used to auto-detect and now you
actually have to tell it?

It also seems to take significantly longer to build
f4pga-arch-defs.xc7a50t for some reason... I didn't see VPR using any
multithreading.