Oops, we do need `time` after all
VTR only checks whether `/usr/bin/env` is available to determine whether
to use `time`, not `time` itself, meaning that on macOS where there's no
sandbox it'll try to use it and fail if it isn't provided.

I decided to patch VTR not to use /usr/bin/env in the first place, so
that enabling/disabling the sandbox doesn't alter behaviour (which would
be a bit gross IMO).