Fix SharedData problems
It seems to be another bug with downloadRecord.dat handling, but not in
our favour this time. So, forgo that entirely by including one
(configurable) device in the base archive test, then deleting it from
the instRecord.dat so it doesn't mess with later tests, and subtracting
the archives of the device we included from `baseArchives` after
installing it again with the modified `instRecord.dat`.

Something I didn't consider when removing Shared Data from the list of
base archives is that there's probably something which does actually
need those archives; and indeed, they're needed for the Versal device
support, so they've now been added there.