Stop tricking pycapnp into thinking it built capnproto
It doesn't work on Linux now that capnproto is built as a shared library
instead of a static one (capnproto doesn't get added to the RUNPATH),
presumably since the bundled one is always built as static.

So, now we patch pycapnp to look in the capnproto derivation for its
`.capnp` files instead of getting it to copy them.

Someone also beat me to the punch on fixing pycapnp in upstream nixpkgs,
so all of that code can be removed.