Initial packaging of Vivado 2025.1
I had to rejigger archive tests again, because using instRecord has
started to become buggy (well, in actuality, the problem is that it
doesn't replicate the bugs found during regular installation) - so, we
have no choice but to use from-scratch installs every time.

However, it turns out Vivado doesn't actually do any checksums on its
archives or after installation, meaning we can pick the smallest archive
of the lot and use it for everything, cutting the time taken by archive
tests down to being nearly negligible (I think it was 16 mins or so).