Remove implied Python deps
For format = "pyproject", wheel is implied, and for format =
"setuptools", both setuptools and wheel are implied.

I only ever tested removing both setuptools and wheel, not just wheel,
which led me to the conclusion that they both had to be specified...
sigh.

I need to do this to pycapnp too at some point, now. I want to check if
`pyproject = true` or `format = "pyproject"` is preferred first, though.