I wonder...
According to https://github.com/macports/macports-ports/pull/23934, the
issue is to do with which version of Xcode the system is using, but for
something running inside Nix it shouldn't make a difference. Would
turning on the sandbox help?

(Wait, maybe this is related to polyml failing to build with the sandbox
enabled?)